src/root-requirements.txt
gdal
geojson-rewind
numpy
Pillow
requests
scipy
